Project_14_Salary_Prediction_using_Machine_Learning.ipynb.ipynb
In this project, we built a supervised machine learning model that accurately predicts a person’s salary based on various features like years of experience, education level, role, and other industry-related attributes. This is a regression problem, and we used popular algorithms like Linear Regression, Decision Trees, and Random Forest Regressor to build and compare models.